Split code into separate files and components
