+++ src/services/capitalize.js
@@ -1,0 +1,3 @@
+export default function capitalize(str) {
+  return str.charAt(0).toUpperCase() + str.slice(1)
+}
