+++ src/services/get-redirect-uri.js
@@ -1,0 +1,3 @@
+export default function getRedirectURI() {
+  return window.location.origin + window.location.pathname + '?authenticated=true';
+}
