+++ src/state/form-context.js
@@ -1,0 +1,5 @@
+import { createContext } from 'react'
+
+const FormContext = createContext({})
+
+export default FormContext
